pyproject.toml
[project]
name = "mcp-debugger"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.0",
    "fastmcp>=2.14.5",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
]
